equiv: tier-10 kfunc prototypes from the kernel's own BTF
Kfunc signatures no longer come from intersecting what the two objects
declare. Each object declares only the kfuncs it calls, in its own BTF,
and they disagree constantly in ways that say nothing about behaviour --
the C object has `struct bpf_dynptr` at 16 bytes where rustc declares the
pointee opaque at 0 -- so the size had to be dropped and the model bailed.
That was the biggest remaining bail class at 102 sites. The kernel has the
real prototype, and being one table handed to both objects it cannot
manufacture a divergence, exactly as with the helper table in tier 9.

222 of the 223 kfuncs the corpus calls resolve from vmlinux plus the five
selftest module BTFs. Unsized kfunc sites: 102 -> 9.

Buffer extents come from the kernel's own kfunc ABI, which turned out to
be recorded in the BTF parameter NAMES (bpfcore was discarding them):
`p__sz` gives the preceding pointer's length, `p__str` marks a
NUL-terminated string and routes to the _cstr_bytes capture the string
kfuncs already used. So bpf_sock_addr_set_sun_path(sa, sun_path,
sun_path__sz) sizes its buffer from the next argument, and
bpf_stream_vprintk(id, fmt__str, args, len__sz) gets both treatments.
Same bail-not-guess rule as the helpers.

ONE MODEL BUG, exposed only because these calls stopped bailing: the
generic kfunc model still compared pointer arguments by raw identity --
the bug already fixed for the generic HELPER model in tier 9.
kfunc_call_test's payloads were identical except the frame offset (the
C's `p1` at stack:T+496 where the translation has it at +376). Private
pointers now compare by aliasing structure, as helpers do.

CORPUS: 1321 EQUIV / 0 INEQUIV / 19 WAIVED, 374 objects fully proved
(from 1274 / 356). Three real translation bugs, all QEMU-gated:

- test_xdp_pull_data narrowed a pointer difference to i32 where C's
  `int data_len != data_end - data` promotes to 64 bits.
- test_sig_in_xattr: the C's set_if_not_errno_or_zero(ret, -EFAULT) means
  "keep a valid errno", but clang holds the kfunc's `int` return via MOV32
  (`w7 = w0`, then `w7 = w7`), which ZERO-extends, and the asm then
  compares the full 64-bit register -- so the errno-preserving branch is
  unreachable in the compiled object and every nonzero return becomes
  -EFAULT. The translation compared as i32 and kept errnos the C object
  throws away; it now mirrors the object's width, disassembly in a comment.
- read_vsyscall waived: its two `*_str` kfuncs have a bare `void *dst`
  that makes add_ksyms.py emit IR this LLVM rejects, so the translation
  reimplements their contract on bpf_probe_read_user_str.

METHOD NOTE: the first sweep ran 12-way and reported test_tunnel_kern as
TIMEOUT with ZERO programs; standalone it takes 24.7s and proves 28/31.
sweep.sh gives each object a fixed 120s while N run in parallel, so load
alone can fabricate a TIMEOUT -- which silently dropped 28 proved programs
AND stopped 4 waivers from firing, the tier-7b failure mode. Re-run at
8-way for the committed numbers. Real regressions: `stream` and
`read_cgroupfs_xattr` went BAIL -> TIMEOUT because the model now goes
deeper (cost, not divergence).

5 new hermetic controls (47 total), each verified to FAIL against the old
behaviour rather than passing vacuously. make ci-local green.

+++ progs/test_xdp_pull_data.rs
@@ -68,7 +68,11 @@ fn do_xdp_pull_data(ctx: *const xdp_md) -> i32 {
     let data_end = vload!((*ctx).data_end) as usize;
     let data = vload!((*ctx).data) as usize;
 
-    if unsafe { data_len } != (data_end - data) as i32 {
+    // C: `if (data_len != data_end - data)`. The pointer difference is
+    // 64-bit and `int data_len` promotes to it, so the comparison is done
+    // at 64 bits; narrowing the difference to i32 first would match on the
+    // low half alone.
+    if unsafe { data_len } as i64 != (data_end - data) as i64 {
         return XDP_DROP;
     }
